Phase 4 Wave A: move the arcade catalog into surfaces/catalog/
Fourth git-mv batch of the core/+surfaces/ scaffolding (admin was #1070,
party #1072, review #1073). ArcadeCatalog and its supporting modules —
CatalogRail, catalogCategory, catalogPagination, catalogScrollIdle,
catalogSort — move to apps/web/src/surfaces/catalog/.

catalog.ts (the games API) and recentPlays.ts stay put: both are imported
well beyond the catalog surface (GamePage, GameTheater, GameDetailPage,
CreatePage, HeroPromptSection, PublishedGameFrame, and more) — shared
primitives, not catalog-specific, same call as leaving mp/protocol.ts and
proposalsApi.ts in place for the party and review moves.

Cross-surface imports repointed: App.tsx's ArcadeCatalog import (still a
static, non-lazy import — the catalog is the default view), CreatePage.tsx's
and surfaces/party/PartyPage.tsx's CatalogRail import.

Baselines resealed for the new paths; stale root-path entries removed.
