+++ apps/web/src/surfaces/catalog/CatalogRail.test.tsx
@@ -4,8 +4,8 @@ import { act, createElement } from 'react';
 import { createRoot, type Root } from 'react-dom/client';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { CatalogRail, FeaturedGame } from './CatalogRail.js';
-import type { CatalogEntry } from './catalog.js';
-import i18n from './i18n/index.js';
+import type { CatalogEntry } from '../../catalog.js';
+import i18n from '../../i18n/index.js';
 
 const withVideo: CatalogEntry = {
   slug: 'neon-courier',
