+++ apps/web/src/surfaces/party/PartyPage.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { CatalogRail } from '../../CatalogRail.js';
+import { CatalogRail } from '../catalog/CatalogRail.js';
 import type { CatalogEntry } from '../../catalog.js';
 import { MascotMoment } from '../../Mascot.js';
 import { PartyDiagram } from './PartyDiagram.js';
