+++ eslint-rules/comment-prose-baseline.json
@@ -373,8 +373,6 @@
     "apps/web/src/AppleSignInButton.tsx": 268,
     "apps/web/src/AppUpdateBanner.test.tsx": 36,
     "apps/web/src/AppUpdateBanner.tsx": 201,
-    "apps/web/src/ArcadeCatalog.test.tsx": 74,
-    "apps/web/src/ArcadeCatalog.tsx": 830,
     "apps/web/src/AuthContext.tsx": 229,
     "apps/web/src/AuthModal.tsx": 79,
     "apps/web/src/betaSplashFit.test.ts": 158,
@@ -384,8 +382,6 @@
     "apps/web/src/BuilderModeBadge.tsx": 80,
     "apps/web/src/catalog.test.ts": 85,
     "apps/web/src/catalog.ts": 535,
-    "apps/web/src/catalogScrollIdle.ts": 142,
-    "apps/web/src/catalogSort.ts": 196,
     "apps/web/src/ClaimHandleModal.tsx": 67,
     "apps/web/src/ClosedBetaSplash.tsx": 146,
     "apps/web/src/CodeSurface.test.tsx": 114,
@@ -577,6 +573,10 @@
     "apps/web/src/surfaces/admin/SuggestionsPanel.tsx": 113,
     "apps/web/src/surfaces/admin/WaitlistPanel.test.tsx": 11,
     "apps/web/src/surfaces/admin/WaitlistPanel.tsx": 164,
+    "apps/web/src/surfaces/catalog/ArcadeCatalog.test.tsx": 74,
+    "apps/web/src/surfaces/catalog/ArcadeCatalog.tsx": 830,
+    "apps/web/src/surfaces/catalog/catalogScrollIdle.ts": 142,
+    "apps/web/src/surfaces/catalog/catalogSort.ts": 196,
     "apps/web/src/surfaces/party/ControllerView.tsx": 127,
     "apps/web/src/surfaces/party/mpApi.ts": 23,
     "apps/web/src/surfaces/party/PartyStage.tsx": 171,
