+++ eslint-rules/module-size-baseline.json
@@ -490,8 +490,6 @@
     "apps/web/src/AppLoadingScreen.tsx": 18,
     "apps/web/src/AppUpdateBanner.test.tsx": 129,
     "apps/web/src/AppUpdateBanner.tsx": 91,
-    "apps/web/src/ArcadeCatalog.test.tsx": 1158,
-    "apps/web/src/ArcadeCatalog.tsx": 1115,
     "apps/web/src/assessmentExportApi.test.ts": 46,
     "apps/web/src/assessmentExportApi.ts": 48,
     "apps/web/src/attachmentImages.ts": 74,
@@ -514,16 +512,6 @@
     "apps/web/src/BuilderModeBadge.tsx": 111,
     "apps/web/src/catalog.test.ts": 247,
     "apps/web/src/catalog.ts": 259,
-    "apps/web/src/catalogCategory.test.ts": 78,
-    "apps/web/src/catalogCategory.ts": 52,
-    "apps/web/src/catalogPagination.test.ts": 25,
-    "apps/web/src/catalogPagination.ts": 19,
-    "apps/web/src/CatalogRail.test.tsx": 211,
-    "apps/web/src/CatalogRail.tsx": 408,
-    "apps/web/src/catalogScrollIdle.test.ts": 69,
-    "apps/web/src/catalogScrollIdle.ts": 92,
-    "apps/web/src/catalogSort.test.ts": 178,
-    "apps/web/src/catalogSort.ts": 244,
     "apps/web/src/ClaimHandleModal.tsx": 185,
     "apps/web/src/ClosedBetaSplash.test.ts": 431,
     "apps/web/src/ClosedBetaSplash.tsx": 181,
@@ -817,6 +805,18 @@
     "apps/web/src/surfaces/admin/SuggestionsPanel.tsx": 146,
     "apps/web/src/surfaces/admin/WaitlistPanel.test.tsx": 197,
     "apps/web/src/surfaces/admin/WaitlistPanel.tsx": 346,
+    "apps/web/src/surfaces/catalog/ArcadeCatalog.test.tsx": 1163,
+    "apps/web/src/surfaces/catalog/ArcadeCatalog.tsx": 1114,
+    "apps/web/src/surfaces/catalog/catalogCategory.test.ts": 77,
+    "apps/web/src/surfaces/catalog/catalogCategory.ts": 52,
+    "apps/web/src/surfaces/catalog/catalogPagination.test.ts": 25,
+    "apps/web/src/surfaces/catalog/catalogPagination.ts": 19,
+    "apps/web/src/surfaces/catalog/CatalogRail.test.tsx": 210,
+    "apps/web/src/surfaces/catalog/CatalogRail.tsx": 407,
+    "apps/web/src/surfaces/catalog/catalogScrollIdle.test.ts": 69,
+    "apps/web/src/surfaces/catalog/catalogScrollIdle.ts": 92,
+    "apps/web/src/surfaces/catalog/catalogSort.test.ts": 177,
+    "apps/web/src/surfaces/catalog/catalogSort.ts": 244,
     "apps/web/src/surfaces/party/ControllerView.test.tsx": 147,
     "apps/web/src/surfaces/party/ControllerView.tsx": 328,
     "apps/web/src/surfaces/party/mpApi.ts": 41,
